refactor: ChatMessageResponse 통일
